+++ src/main/java/roomescape/reservation/dto/MemberReservationAddRequest.java
@@ -3,6 +3,7 @@ package roomescape.reservation.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import roomescape.payment.dto.PaymentConfirmRequest;
 
@@ -12,7 +13,7 @@ public record MemberReservationAddRequest(
         @NotNull(message = "테마 선택은 필수 입니다.") @Positive Long themeId,
         @NotBlank String paymentKey,
         @NotBlank String orderId,
-        @NotNull @Positive Integer amount) {
+        @NotNull @Positive BigDecimal amount) {
 
     public PaymentConfirmRequest extractPaymentInformation() {
         return new PaymentConfirmRequest(paymentKey, orderId, amount);
